revno: 13541 message:   Fixed test error
committer: Lars Helge Øverland <larshelge@gmail.com>
branch nick: dhis2
timestamp: Thu 2014-01-02 15:21:19 +0100

+++ dhis-2/dhis-services/dhis-service-patient/src/main/java/org/hisp/dhis/program/ProgramInstanceDeletionHandler.java
@@ -123,14 +123,17 @@ public class ProgramInstanceDeletionHandler
                 patientCommentService.deletePatientComment( patientComment );
             }
 
-            for ( PatientAttribute patientAttribute : programInstance.getProgram().getPatientAttributes() )
-            {
-                PatientAttributeValue patientAttributeValue = patientAttributeValueService.getPatientAttributeValue(
-                    patient, patientAttribute );
-
-                if ( patientAttributeValue != null )
+            if ( programInstance.getProgram() != null && programInstance.getProgram().getPatientAttributes() != null )
+            {            
+                for ( PatientAttribute patientAttribute : programInstance.getProgram().getPatientAttributes() )
                 {
-                    patientAttributeValueService.deletePatientAttributeValue( patientAttributeValue );
+                    PatientAttributeValue patientAttributeValue = patientAttributeValueService.getPatientAttributeValue(
+                        patient, patientAttribute );
+    
+                    if ( patientAttributeValue != null )
+                    {
+                        patientAttributeValueService.deletePatientAttributeValue( patientAttributeValue );
+                    }
                 }
             }
